imagePullPolicy always for steps with non-pinned images
Tekton pods default to the IfNotPresent imagePullPolicy, leading to the possibility of missing updates to the images when they are referred by tag and already pulled by a past execution in the node where a TaskRun pod is scheduled.

This commit sets the imagePullPolicy to Always for the steps using non-pinned images referenced by tag.